fixed sending wrong data type

+++ src/arduino_communicator/scripts/arduino_communicator_node.py
@@ -109,14 +109,14 @@ class RoverController:
         command_struct = [0]*8
 
         # Scales commands to be between [0,200]
-        command_struct[0] = 100 + (100 * lin_vel)
-        command_struct[1] = 100 + (100 * ang_vel)
-        command_struct[2] = 100 + (100 * base_rot)
-        command_struct[3] = 100 + (100 * arm_act_1)
-        command_struct[4] = 100 + (100 * arm_act_2)
-        command_struct[5] = 100 + (100 * wrist_rot)
-        command_struct[6] = 100 + (100 * wrist_act)
-        command_struct[7] = 100 + (100 * gripper)
+        command_struct[0] = int(100 + (100 * lin_vel))
+        command_struct[1] = int(100 + (100 * ang_vel))
+        command_struct[2] = int(100 + (100 * base_rot))
+        command_struct[3] = int(100 + (100 * arm_act_1))
+        command_struct[4] = int(100 + (100 * arm_act_2))
+        command_struct[5] = int(100 + (100 * wrist_rot))
+        command_struct[6] = int(100 + (100 * wrist_act))
+        command_struct[7] = int(100 + (100 * gripper))
 
         rospy.loginfo("Commands of type %s", type(command_struct[0]))
